+++ R/early_day_motions.R
@@ -1,6 +1,8 @@
 
 
-#' Imports data on early day motions, including the content, signatories, and sponsors of early day motions.
+#' Imports data on early day motions
+#'
+#' Includes the content, signatories, and sponsors of early day motions.
 #' @param edm_id Accepts the ID number of an early day motion, and returns data on that motion. Note that EDM numbers reset each parliamentary session, so using this as the only parameter will return multiple early day motions with the same ID code. If NULL, returns all available Early Day Motions. Note that there, are as of 2017-06-15, 43330 early day motions on listed in the API, so requesting all early day motions without other parameters is slow and very demanding on the API itself. Defaults to NULL.
 #' @param session Accepts a parliamentary session, in 'yyyy/yy' format. Defaults to NULL.
 #' @param start_date The earliest date to include in the tibble. Defaults to '1900-01-01'. Accepts character values in 'YYYY-MM-DD' format, and objects of class Date, POSIXt, POSIXct, POSIXlt or anything else than can be coerced to a date with \code{as.Date()}.