+++ R/hansard-package.R
@@ -78,6 +78,4 @@
 #' @import tibble
 #' @import tidyr
 #' @importFrom lubridate parse_date_time
-#' @useDynLib hansard
-#' @importFrom Rcpp sourceCpp
 NULL